+++ publish.sh
@@ -4,13 +4,14 @@ Script to publish a new version of this library on PyPI
 
 Args:
     # These environment variables must / should be set
-    GITHUB_USERNAME : username for pypi
+    TWINE_USERNAME : username for pypi
     TWINE_PASSWORD : password for pypi
     USE_GPG : defaults to True
-    GPG_IDENTIFIER
 
 Requirements:
-     twine
+     twine >= 1.13.0
+     gpg2 >= 2.2.4
+     OpenSSL >= 1.1.1c
 
 Notes:
     # NEW API TO UPLOAD TO PYPI
@@ -20,57 +21,115 @@ Usage:
     cd <YOUR REPO>
 
     # Set your variables or load your secrets
-    export GITHUB_USERNAME=<pypi-username>
+    export TWINE_USERNAME=<pypi-username>
     export TWINE_PASSWORD=<pypi-password>
 
     source $(secret_loader.sh)
 
-    # Interactive/Dry run
-    ./publish.sh 
+    MB_PYTHON_TAG=cp37-cp37m 
+    MB_PYTHON_TAG=cp36-cp36m 
+    MB_PYTHON_TAG=cp35-cp35m 
+    MB_PYTHON_TAG=cp27-cp27mu
 
-    # Non-Interactive run
-    #./publish.sh yes
+    echo "MB_PYTHON_TAG = $MB_PYTHON_TAG"
+    MB_PYTHON_TAG=$MB_PYTHON_TAG ./run_multibuild.sh
+    DEPLOY_BRANCH=master DEPLOY_REMOTE=ibeis MB_PYTHON_TAG=$MB_PYTHON_TAG ./publish.sh yes
 '''
 
+check_variable(){
+    KEY=$1
+    VAL=${!KEY}
+    echo "[DEBUG] CHECK VARIABLE: $KEY=\"$VAL\""
+    if [[ "$VAL" == "" ]]; then
+        echo "[ERROR] UNSET VARIABLE: $KEY=\"$VAL\""
+        exit 1;
+    fi
+}
+
 # Options
-if [[ "$USER" == "joncrall" ]]; then
-    GITHUB_USERNAME=erotemic
-fi
-TWINE_PASSWORD=${TWINE_PASSWORD:="<unknown>"}
+CURRENT_BRANCH=${CURRENT_BRANCH:=$(git branch | grep \* | cut -d ' ' -f2)}
+DEPLOY_BRANCH=${DEPLOY_BRANCH:=release}
+DEPLOY_REMOTE=${DEPLOY_REMOTE:=origin}
+NAME=${NAME:=$(python -c "import setup; print(setup.NAME)")}
+VERSION=$(python -c "import setup; print(setup.VERSION)")
+MB_PYTHON_TAG=${MB_PYTHON_TAG:=$(python -c "import setup; print(setup.MB_PYTHON_TAG)")}
+
+check_variable CURRENT_BRANCH
+check_variable DEPLOY_BRANCH
+check_variable DEPLOY_REMOTE
+check_variable VERSION || exit 1
+
 TAG_AND_UPLOAD=${TAG_AND_UPLOAD:=$1}
+TWINE_USERNAME=${TWINE_USERNAME:=""}
+TWINE_PASSWORD=${TWINE_PASSWORD:=""}
+
 USE_GPG=${USE_GPG:="True"}
 
-# First tag the source-code
-TRAVIS_BRANCH=${TRAVIS_BRANCH:=$(git branch | grep \* | cut -d ' ' -f2)}
-DEPLOY_BRANCH=${DEPLOY_BRANCH:=release}
-VERSION=$(python -c "import setup; print(setup.parse_version())")
-GPG_EXECUTABLE=${GPG_EXECUTABLE:=gpg}
+if [[ "$(which gpg2)" != "" ]]; then
+    GPG_EXECUTABLE=${GPG_EXECUTABLE:=gpg2}
+else
+    GPG_EXECUTABLE=${GPG_EXECUTABLE:=gpg}
+fi
+
+GPG_KEYID=${GPG_KEYID:=$(git config --local user.signingkey)}
+GPG_KEYID=${GPG_KEYID:=$(git config --global user.signingkey)}
 
 
 echo "
 === PYPI BUILDING SCRIPT ==
-TRAVIS_BRANCH='$TRAVIS_BRANCH'
+CURRENT_BRANCH='$CURRENT_BRANCH'
 DEPLOY_BRANCH='$DEPLOY_BRANCH'
 VERSION='$VERSION'
-GITHUB_USERNAME='$GITHUB_USERNAME'
+TWINE_USERNAME='$TWINE_USERNAME'
+GPG_KEYID = '$GPG_KEYID'
+MB_PYTHON_TAG = '$MB_PYTHON_TAG'
 "
 
 
 echo "
 === <BUILD WHEEL> ===
 "
+
+
+
 echo "LIVE BUILDING"
 # Build wheel and source distribution
-python setup.py bdist_wheel --universal
-python setup.py sdist 
 
-BDIST_WHEEL_PATH=$(ls dist/*-$VERSION-*.whl)
-SDIST_PATH=$(dir dist/*-$VERSION*.tar.gz)
+MODE=${MODE:=all}
+
+WHEEL_PATHS=()
+
+if [[ "$MODE" == "sdist" ]]; then
+    python setup.py sdist 
+    WHEEL_PATH=$(ls dist/$NAME-$VERSION*.tar.gz)
+    WHEEL_PATHS+=($WHEEL_PATH)
+elif [[ "$MODE" == "universal" ]]; then
+    python setup.py bdist_wheel --universal
+    WHEEL_PATH=$(ls dist/*-$VERSION-$MB_PYTHON_TAG*.whl)
+    WHEEL_PATHS+=($WHEEL_PATH)
+elif [[ "$MODE" == "bdist" ]]; then
+    echo "Assume wheel has already been built"
+    WHEEL_PATH=$(ls wheelhouse/*-$VERSION-$MB_PYTHON_TAG*.whl)
+    WHEEL_PATHS+=($WHEEL_PATH)
+elif [[ "$MODE" == "all" ]]; then
+    python setup.py sdist 
+    WHEEL_PATH=$(ls dist/$NAME-$VERSION*.tar.gz)
+    WHEEL_PATHS+=($WHEEL_PATH)
+
+    python setup.py bdist_wheel --universal
+    WHEEL_PATH=$(ls dist/*-$VERSION-$MB_PYTHON_TAG*.whl)
+    WHEEL_PATHS+=($WHEEL_PATH)
+else
+    echo "bad mode"
+    exit 1
+fi
+
 echo "
-echo "VERSION='$VERSION'"
-BDIST_WHEEL_PATH='$BDIST_WHEEL_PATH'
-SDIST_PATH='$SDIST_PATH'
+MODE=$MODE
+VERSION='$VERSION'
+WHEEL_PATHS='$WHEEL_PATHS'
 "
+
 echo "
 === <END BUILD WHEEL> ===
 "
@@ -78,91 +137,91 @@ echo "
 echo "
 === <GPG SIGN> ===
 "
-if [ "$USE_GPG" == "True" ]; then
-    # https://stackoverflow.com/questions/45188811/how-to-gpg-sign-a-file-that-is-built-by-travis-ci
-    # secure gpg --export-secret-keys > all.gpg
 
-    #GPG_IDENTIFIER=${GPG_IDENTIFIER:-"travis-ci-Erotemic"}
-    #GPG_KEYID=$(gpg --list-keys --keyid-format LONG "$GPG_IDENTIFIER" | head -n 2 | tail -n 1 | awk '{print $1}' | tail -c 9)
-    GPG_KEYID=D297D757
-    # REQUIRES GPG >= 2.2
-    echo "GPG_IDENTIFIER = $GPG_IDENTIFIER"
-    echo "GPG_KEYID=$GPG_KEYID"
 
-    echo "Removing old signatures"
-    rm dist/*.asc
+for WHEEL_PATH in "${WHEEL_PATHS[@]}"
+do
+    echo "WHEEL_PATH = $WHEEL_PATH"
+    check_variable WHEEL_PATH
+    if [ "$USE_GPG" == "True" ]; then
+        # https://stackoverflow.com/questions/45188811/how-to-gpg-sign-a-file-that-is-built-by-travis-ci
+        # secure gpg --export-secret-keys > all.gpg
 
-    echo "Signing wheels"
-    GPG_SIGN_CMD="$GPG_EXECUTABLE --batch --yes --detach-sign --armor --local-user $GPG_KEYID"
-    $GPG_SIGN_CMD --output $BDIST_WHEEL_PATH.asc $BDIST_WHEEL_PATH
-    $GPG_SIGN_CMD --output $SDIST_PATH.asc $SDIST_PATH
+        # REQUIRES GPG >= 2.2
+        check_variable GPG_EXECUTABLE
+        check_variable GPG_KEYID
 
-    echo "Checking wheels"
-    twine check $BDIST_WHEEL_PATH.asc $BDIST_WHEEL_PATH
-    twine check $SDIST_PATH.asc $SDIST_PATH
+        echo "Signing wheels"
+        GPG_SIGN_CMD="$GPG_EXECUTABLE --batch --yes --detach-sign --armor --local-user $GPG_KEYID"
+        echo "GPG_SIGN_CMD = $GPG_SIGN_CMD"
+        $GPG_SIGN_CMD --output $WHEEL_PATH.asc $WHEEL_PATH
 
-    echo "Verifying wheels"
-    gpg --verify $BDIST_WHEEL_PATH.asc $BDIST_WHEEL_PATH 
-    gpg --verify $SDIST_PATH.asc $SDIST_PATH 
-else
-    echo "USE_GPG=False, Skipping GPG sign"
-fi
+        echo "Checking wheels"
+        twine check $WHEEL_PATH.asc $WHEEL_PATH
+
+        echo "Verifying wheels"
+        $GPG_EXECUTABLE --verify $WHEEL_PATH.asc $WHEEL_PATH 
+    else
+        echo "USE_GPG=False, Skipping GPG sign"
+    fi
+done
 echo "
 === <END GPG SIGN> ===
 "
+
+if [[ "$CURRENT_BRANCH" != "$DEPLOY_BRANCH" ]]; then
+    TAG_AND_UPLOAD="no"
+    echo "current branch is not the deploy branch. Forcing tag_and_upload=no"
+fi
 
 
 # Verify that we want to publish
 if [[ "$TAG_AND_UPLOAD" != "yes" ]]; then
     if [[ "$TAG_AND_UPLOAD" != "no" ]]; then
-        read -p "Are you ready to publish version='$VERSION' on branch='$TRAVIS_BRANCH'? (input 'yes' to confirm)" ANS
+        read -p "Are you ready to publish version='$VERSION' on branch='$CURRENT_BRANCH'? (input 'yes' to confirm)" ANS
         echo "ANS = $ANS"
         TAG_AND_UPLOAD="$ANS"
     else
-        echo "Ready to publish VERSION='$VERSION' on branch='$TRAVIS_BRANCH'" 
+        echo "WRONG BRANCH: Not ready to publish VERSION='$VERSION' on branch='$CURRENT_BRANCH'" 
     fi
 else
-    echo "Not ready to publish VERSION='$VERSION' on branch='$TRAVIS_BRANCH'" 
+    echo "Do not want to publish VERSION='$VERSION' on branch='$CURRENT_BRANCH'" 
 fi
+
 
 if [[ "$TAG_AND_UPLOAD" == "yes" ]]; then
+    check_variable TWINE_USERNAME
+    check_variable TWINE_PASSWORD
 
-    if [[ "$GITHUB_USERNAME" == "" ]]; then
-        echo "Error GITHUB_USERNAME is not set"
-        exit 1
-    fi
-    if [[ "$TWINE_PASSWORD" == "" ]]; then
-        echo "Error TWINE_PASSWORD is not set"
-        exit 1
-    fi
+    #git tag $VERSION -m "tarball tag $VERSION"
+    #git push --tags $DEPLOY_REMOTE $DEPLOY_BRANCH
 
-    if [[ "$TRAVIS_BRANCH" == "$DEPLOY_BRANCH" ]]; then
-        echo "TRAVIS_BRANCH = $TRAVIS_BRANCH"
-        git tag $VERSION -m "tarball tag $VERSION"
-        git push --tags origin $DEPLOY_BRANCH
+    for WHEEL_PATH in "${WHEEL_PATHS[@]}"
+    do
         if [ "$USE_GPG" == "True" ]; then
-            twine upload --username $GITHUB_USERNAME --password $TWINE_PASSWORD --sign $BDIST_WHEEL_PATH.asc $BDIST_WHEEL_PATH
-            twine upload --username $GITHUB_USERNAME --password $TWINE_PASSWORD --sign $SDIST_PATH.asc $SDIST_PATH
+            twine upload --username $TWINE_USERNAME --password=$TWINE_PASSWORD --sign $WHEEL_PATH.asc $WHEEL_PATH
         else
-            twine upload --username $GITHUB_USERNAME --password $TWINE_PASSWORD $BDIST_WHEEL_PATH 
-            twine upload --username $GITHUB_USERNAME --password $TWINE_PASSWORD $SDIST_PATH 
+            twine upload --username $TWINE_USERNAME --password=$TWINE_PASSWORD $WHEEL_PATH 
         fi
-    else
-        echo "TRAVIS_BRANCH!=DEPLOY_BRANCH. skipping tag and upload"
-        echo "ONLY ABLE TO PUBLISH ON DEPLOY TRAVIS_BRANCH
+    done
+    echo """
+        !!! FINISH: LIVE RUN !!!
+    """
+else
+    echo """
+        DRY RUN ... Skiping tag and upload
 
-        TRAVIS_BRANCH = $TRAVIS_BRANCH
-        DEPLOY_BRANCH = $DEPLOY_BRANCH
-        "
-    fi
-else  
-    echo "Dry run"
-    echo "skiping tag and upload"
+        NAME = '$NAME'
+        VERSION = '$VERSION'
+        DEPLOY_REMOTE = '$DEPLOY_REMOTE'
+        CURRENT_BRANCH = '$CURRENT_BRANCH'
+        DEPLOY_BRANCH = '$DEPLOY_BRANCH'
+        TAG_AND_UPLOAD = '$TAG_AND_UPLOAD'
+        WHEEL_PATH = '$WHEEL_PATH'
+        WHEEL_PATHS = '$WHEEL_PATHS'
+
+        To do live run set TAG_AND_UPLOAD=yes and ensure deploy and current branch are the same
+
+        !!! FINISH: DRY RUN !!!
+    """
 fi
-
-__notes__="""
-Notes:
-    # References: https://docs.travis-ci.com/user/deployment/pypi/
-    travis encrypt TWINE_PASSWORD=$TWINE_PASSWORD  
-    travis encrypt GITHUB_USERNAME=$GITHUB_USERNAME 
-"""