+++ xdoctest/utils/util_str.py
@@ -71,8 +71,17 @@ def color_text(text, color):
 
         if sys.platform.startswith('win32'):  # nocover
             # Hack on win32 to support colored output
-            import colorama
-            colorama.init()
+            try:
+                import colorama
+                colorama.init()
+            except ImportError as ex:
+                import warnings
+                warnings.warn('os is win32 and colorma is not installed {!r}'.format(ex))
+                pass
+            import os
+            if os.environ.get('XDOC_WIN32_COLORS', 'False') == 'False':
+                # hack: dont color on windows by default, but do init colorama
+                return text
 
         try:
             ansi_text = pygments.console.colorize(color, text)