+++ CHANGELOG.md
@@ -5,6 +5,14 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Version 0.11.0] - Unreleased
+
+### Added
+* Add support for PyPy
+* REQUIRES directive now supports CPython, IronPython, Jython, and PyPy
+* REQUIRES directive now supports PY2, PY3
+
+
 ## [Version 0.10.3] - Unreleased
 
 ### Fixed